% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrado_coluna.R
\name{quadrado_coluna}
\alias{quadrado_coluna}
\title{Transformação de coluna elevando ao quadrado}
\usage{
quadrado_coluna(df, nome_coluna)
}
\arguments{
\item{df}{Um objeto do tipo \code{Data.Frame} a ser tranformado.}

\item{nome_coluna}{String identificando o nome da coluna a ter
seu valor transformado. Precisa ser o nome de uma coluna de \code{df}
que tenha os tipos \code{numeric} ou \code{integer}.}
}
\value{
Retorna um objeto do tipo \code{Data.Frame} com os mesmos nomes
de coluna, porém com a coluna especificada substituída por seus valores
ao quadrado.
}
\description{
Esta função substitui uma coluna numérica (\code{numeric} ou \code{integer})
especificada de uma tabela por seu valor ao quadrado.
}
\examples{
quadrado_coluna(iris, "Sepal.Length")
quadrado_coluna(mtcars, "qsec")

}
